re-implement DataContextManager as Class
