+++ shesha-reactjs/src/providers/dataContextManager/models.ts
@@ -1,4 +1,4 @@
-import { DataTypes } from "@/interfaces";
+import { ConfigurableFormInstance, DataTypes } from "@/interfaces";
 import { IContextMetadata } from "@/interfaces/metadata";
 import { DataContextType, IDataContextProviderActionsContext } from "../dataContextProvider/contexts";
 
@@ -39,3 +39,39 @@ export const DEFAULT_CONTEXT_METADATA = {
   properties: [],
   methods: [],
 } as IContextMetadata;
+
+export interface IDataContextManagerState {
+  lastUpdate: number;
+  id: string;
+  parent?: IDataContextManagerFullInstance | undefined;
+
+}
+
+export interface IDataContextsData {
+  [key: string]: unknown;
+  lastUpdate: number;
+}
+
+export interface IDataContextManagerActions {
+  registerDataManager: (payload: IDataContextManagerFullInstance) => void;
+  unregisterDataManager: (payload: IDataContextManagerFullInstance) => void;
+  registerDataContext: (payload: IRegisterDataContextPayload) => void;
+  unregisterDataContext: (payload: IRegisterDataContextPayload) => void;
+  getLocalDataContexts: (contextId: string) => IDataContextDescriptor[];
+  getDataContexts: (topId?: string) => IDataContextDescriptor[];
+  getDataContextsData: (topId?: string, data?: IDataContextsData) => IDataContextsData;
+  getDataContext: (contextId: string) => IDataContextDescriptor | undefined;
+  getNearestDataContext: (topId: string, type: DataContextType) => IDataContextDescriptor | undefined;
+  getDataContextData: (contextId: string) => unknown;
+  onChangeContext: (dataContext: IDataContextDescriptor) => void;
+  onChangeContextData: () => void;
+  updatePageFormInstance: (form: ConfigurableFormInstance) => void;
+  getPageFormInstance: () => ConfigurableFormInstance | undefined;
+  getPageContext: () => IDataContextDescriptor | undefined;
+  getRoot: () => IDataContextManagerFullInstance | undefined;
+  getParent: () => IDataContextManagerFullInstance | undefined;
+}
+
+export interface IDataContextManagerFullInstance extends IDataContextManagerState, IDataContextManagerActions {
+
+}
