+++ shesha-reactjs/src/providers/dynamicActions/implementations/dataSourceDynamicMenu/entityDynamicMenuItem/index.tsx
@@ -3,7 +3,6 @@ import { ButtonGroupItemProps, IButtonGroupItem } from '@/providers/buttonGroupC
 import {
   DynamicActionsProvider,
   DynamicItemsEvaluationHook,
-  useDataContextManagerActionsOrUndefined,
   useFormData,
   useGlobalState,
   useHttpClient,
@@ -21,6 +20,7 @@ import { isDefined, isNullOrWhiteSpace } from '@/utils/nullables';
 import { getEntityTypeIdentifierQueryParams, isEntityTypeIdentifier } from '@/providers/metadataDispatcher/entities/utils';
 import { buildUrl } from '@/utils/url';
 import { convertDotNotationPropertiesToGraphQL } from '@/providers/form/utils';
+import { useDataContextManagerActionsOrUndefined } from '@/providers/dataContextManager/hooks';
 
 type ArrayOrObjectWithItems<T> = T[] | {
   items: T[];
