+++ shesha-reactjs/src/providers/form/utils.ts
@@ -15,6 +15,7 @@ import {
   isConfigurableFormComponent,
   isObjectWithStringId,
   isRawComponentsContainer,
+  RootContexts,
   STYLE_BOX_CSS_POPERTIES,
   StyleBoxValue,
   useDataTableStateOrUndefined,
@@ -23,13 +24,9 @@ import {
   useMetadataDispatcher,
 } from '@/providers';
 import {
-  IDataContextManagerActionsContext,
-  IDataContextManagerFullInstance,
-  IDataContextsData,
-  RootContexts,
   useDataContextManagerActionsOrUndefined,
   useDataContextManagerOrUndefined,
-} from '@/providers/dataContextManager';
+} from '@/providers/dataContextManager/hooks';
 import { IDataContextFull, useDataContextOrUndefined } from '@/providers/dataContextProvider/contexts';
 import { ISelectionProps } from '@/providers/dataTable/interfaces';
 import { executeFunction } from '@/utils';
@@ -50,7 +47,7 @@ import moment from 'moment';
 import Mustache from 'mustache';
 import { CSSProperties, useRef } from 'react';
 import { IArgumentsEvaluationContext } from '../configurableActionsDispatcher/contexts';
-import { SheshaCommonContexts } from '../dataContextManager/models';
+import { IDataContextManagerActions, IDataContextManagerFullInstance, IDataContextsData, SheshaCommonContexts } from '../dataContextManager/models';
 import { GetShaFormDataAccessor } from '../dataContextProvider/contexts/shaDataAccessProxy';
 import { ISetStatePayload } from '../globalState/contexts';
 import { IParentProviderProps, useParentOrUndefined } from '../parentProvider/index';
@@ -171,7 +168,7 @@ export type AvailableConstantsContext = {
   componentApi?: IComponentApiActions | undefined;
   closestShaFormApi: IFormApi | undefined;
   selectedRow?: ISelectionProps | undefined;
-  dcm: IDataContextManagerActionsContext | undefined;
+  dcm: IDataContextManagerActions | undefined;
   metadataDispatcher: IMetadataDispatcher | undefined;
   closestContextId: string | undefined;
   globalState: IAnyObject | undefined;
